+++ winch/codegen/src/isa/mod.rs
@@ -1,6 +1,7 @@
 use anyhow::{anyhow, Result};
 use core::fmt::Formatter;
 use cranelift_codegen::isa::{CallConv, IsaBuilder};
+use cranelift_codegen::settings;
 use cranelift_codegen::{Final, MachBufferFinalized};
 use std::{
     error,
@@ -75,6 +76,12 @@ pub trait TargetIsa: Send + Sync {
 
     /// Get the target triple of the ISA.
     fn triple(&self) -> &Triple;
+
+    /// Get the ISA-independent flags that were used to make this trait object.
+    fn flags(&self) -> &settings::Flags;
+
+    /// Get the ISA-dependent flag values that were used to make this trait object.
+    fn isa_flags(&self) -> Vec<settings::Value>;
 
     fn compile_function(
         &self,
